requirements.txt
h5py
hdf5plugin
pandas
numpy
scipy
matplotlib
pyyaml
mzapy
polars
